requirements.txt
pandas
numpy
kafka-python
requests